added fuel motorcycle and truck

+++ Ex03.GarageLogic/Question.cs
@@ -3,11 +3,11 @@ namespace Ex03.GarageLogic;
 public class Question
 {
     public string m_Question_text { get; set; }
-    public int m_Min { get; set; }
-    public int m_Max { get; set; }
+    public float m_Min { get; set; }
+    public float m_Max { get; set; }
     public Type m_Type { get; set; }
 
-    public Question(string i_Question_text, Type i_Type, int i_Min = 0, int i_Max = 0)
+    public Question(string i_Question_text, Type i_Type, float i_Min = 0, float i_Max = 0)
     {
         m_Question_text = i_Question_text;
         m_Min = i_Min;
